Implement Alex's simple PR comments
